pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 61"]

[project]
authors = [{name = "Andrea Zoppi", email = "texzk@email.it"}]
classifiers = [
    # complete classifier list:
    # http://pypi.python.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
    "Topic :: Software Development",
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: Utilities",
]
dependencies = [
    "bytesparse >= 0.0.8",  # comment out and use the one below for development instead
    # "bytesparse @ git+https://github.com/TexZK/bytesparse.git",  # only for development
    "click",
    "colorama",
]
description = "Library to handle hexadecimal record files"
dynamic = ["version"]
license = {text = "BSD-2-Clause"}
name = "hexrec"
readme = "README.rst"
requires-python = ">= 3.7"

[project.optional-dependencies]
testing = ["coverage", "pyelftools", "pytest"]

[project.scripts]
hexrec = "hexrec.cli:main"

[project.urls]
documentation = "https://hexrec.readthedocs.io"
source = "https://github.com/TexZK/hexrec"
tracker = "https://github.com/TexZK/hexrec/issues"

[tool.setuptools.dynamic]
version = {attr = "hexrec.__version__"}
